# uk-mrtm-app-api/src/main/resources/openapi/uk-maritime-ext-openapi-1.2.0.json
{
  "openapi": "3.0.1",
  "info":
  {
    "title": "Maritime External API",
    "description": "Maritime external API description.",
    "version": "1.2.0"
  },
  "servers":
  [
    {
      "url": "https://manage-emissions-reporting.service.gov.uk/maritime/api",
      "description": "Production environment"
    }
  ],
  "security":
  [
    {
      "bearerAuth":
      []
    }
  ],
  "paths":
...
    "/external/v1.0/accounts/{company-imo-number}/emp":
    {
      "put":
      {
        "tags":
        [
          "Maritime emissions monitoring plan API"
        ],
        "summary": "Saves or updates maritime emissions monitoring plan data",
        "operationId": "submitEmissionsMonitoringPlanData",
        "parameters":
        [
          {
            "name": "company-imo-number",
            "in": "path",
            "description": "The company IMO number",
            "required": true,
            "schema":
            {
              "pattern": "^\\d{7}$",
              "type": "string"
            }
          }
        ],
        "requestBody":
        {
          "content":
          {
            "application/json":
            {
              "schema":
              {
                "$ref": "#/components/schemas/ExternalEmissionsMonitoringPlan"
              }
            }
          },
          "required": true
        },
        "responses":
        {
          "404":
          {
            "description": "Not Found",
            "content":
            {
              "application/json":
              {
                "schema":
                {
                  "$ref": "#/components/schemas/ExternalErrorResponse"
                }
              }
            }
          },
          "400":
          {
            "description": "Bad Request\t\n\t\n Error Code | Description \t\nFORM1001 | Form validation failed \t\nFORM1002 | Parameters validation failed \t\nEMP1001 | Business validation error",
            "content":
            {
              "application/json":
              {
                "schema":
                {
                  "$ref": "#/components/schemas/ExternalErrorResponse"
                }
              }
            }
          },
          "403":
          {
            "description": "Forbidden",
            "content":
            {
              "application/json":
              {
                "schema":
                {
                  "$ref": "#/components/schemas/ExternalErrorResponse"
                }
              }
            }
          },
          "204":
          {
            "description": "No Content"
          },
          "500":
          {
            "description": "Internal Server Error",
            "content":
            {
              "application/json":
              {
                "schema":
                {
                  "$ref": "#/components/schemas/ExternalErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components":
...
    "schemas":
...
      "ExternalEmpDelegatedResponsibility":
      {
        "required":
        [
          "delegatedResponsibilityUsed"
        ],
        "type": "object",
        "properties":
        {
          "delegatedResponsibilityUsed":
          {
            "type": "boolean",
            "description": "Delegation of UK ETS compliance by registered ship owners. If true, registeredOwners and responsibilityDeclaration are required otherwise they must be omitted"
          },
          "registeredOwners":
          {
            "uniqueItems": true,
            "type": "array",
            "items":
            {
              "$ref": "#/components/schemas/ExternalEmpRegisteredOwner"
            }
          },
          "responsibilityDeclaration":
          {
            "type": "boolean",
            "description": "Maritime Operator declares that the information provided is true. If delegatedResponsibilityUsed is true it must be true, otherwise it must be omitted"
          }
        },
        "description": "Delegated UK ETS responsibility"
      },
      "ExternalErrorResponse":
      {
        "type": "object",
        "properties":
        {
          "code":
          {
            "type": "string"
          },
          "message":
          {
            "type": "string"
          },
          "data":
          {
            "type": "array",
            "items":
            {
              "$ref": "#/components/schemas/Violation"
            }
          }
        }
      },
      "ExternalEmpEmissionsProcedures":
      {
        "required":
        [
          "emissionFactorsProcedure",
          "emissionSourcesProcedure",
          "reductionClaimProcedure"
        ],
        "type": "object",
        "properties":
        {
          "emissionSourcesProcedure":
          {
            "$ref": "#/components/schemas/ExternalEmpProcedureForm"
          },
          "emissionFactorsProcedure":
          {
            "$ref": "#/components/schemas/ExternalEmpEmissionFactorsProcedure"
          },
          "reductionClaimProcedure":
          {
            "$ref": "#/components/schemas/ExternalEmpReductionClaimProcedure"
          }
        },
        "description": "Procedures related to emissions sources and emissions factors"
      },
      "ExternalEmpShipDetails":
      {
        "required":
        [
          "companyNature",
          "flag",
          "grossTonnage",
          "iceClassPolarCode",
          "name",
          "shipImoNumber",
          "shipType"
        ],
        "type": "object",
        "properties":
        {
          "shipImoNumber":
          {
            "pattern": "^\\d{7}$",
            "type": "string"
          },
          "name":
          {
            "maxLength": 255,
            "minLength": 1,
            "type": "string",
            "description": "Name of the Ship as stated on IMO GISIS"
          },
          "shipType":
          {
            "type": "string",
            "enum":
            [
              "PAX",
              "RORO",
              "CONT",
              "OIL",
              "CHEM",
              "LNG",
              "GAS",
              "BULK",
              "GENERAL",
              "RCV",
              "VEH",
              "COMB",
              "ROPAX",
              "CONT_RORO",
              "CRUISE",
              "OFFSHORE",
              "OTHER"
            ]
          },
          "grossTonnage":
          {
            "maximum": 999999999,
            "minimum": 5000,
            "type": "integer",
            "format": "int32"
          },
          "flag":
          {
            "type": "string",
            "description": "Code of the flag",
            "enum":
            [
              "AE",
              "AF",
              "AG",
              "AI",
              "AL",
              "AM",
              "AO",
              "AR",
              "AS",
              "AT",
              "AU",
              "AW",
              "AZ",
              "BA",
              "BB",
              "BD",
              "BE",
              "BG",
              "BH",
              "BJ",
              "BM",
              "BN",
              "BO",
              "BQ",
              "BR",
              "BS",
              "BY",
              "BZ",
              "CA",
              "CD",
              "CG",
              "CH",
              "CI",
              "CK",
              "CL",
              "CM",
              "CN",
              "CO",
              "CR",
              "CU",
              "CV",
              "CW",
              "CY",
              "CZ",
              "DE",
              "DJ",
              "DK",
              "DM",
              "DO",
              "DZ",
              "EC",
              "EE",
              "EG",
              "ER",
              "ES",
              "ET",
              "FI",
              "FJ",
              "FK",
              "FM",
              "FO",
              "FR",
              "GA",
              "GB",
              "GD",
              "GE",
              "GG",
              "GH",
              "GI",
              "GM",
              "GN",
              "GQ",
              "GR",
              "GT",
              "GW",
              "GY",
              "HK",
              "HN",
              "HR",
              "HT",
              "HU",
              "ID",
              "IE",
              "IL",
              "IM",
              "IN",
              "IQ",
              "IR",
              "IS",
              "IT",
              "JE",
              "JM",
              "JO",
              "JP",
              "KE",
              "KG",
              "KH",
              "KI",
              "KM",
              "KN",
              "KP",
              "KR",
              "KW",
              "KY",
              "KZ",
              "LA",
              "LB",
              "LC",
              "LK",
              "LR",
              "LT",
              "LU",
              "LV",
              "LY",
              "MA",
              "MC",
              "MD",
              "ME",
              "MG",
              "MH",
              "ML",
              "MM",
              "MN",
              "MO",
              "MR",
              "MS",
              "MT",
              "MU",
              "MV",
              "MX",
              "MY",
              "MZ",
              "NA",
              "NE",
              "NG",
              "NI",
              "NL",
              "NO",
              "NR",
              "NZ",
              "OM",
              "PA",
              "PE",
              "PF",
              "PG",
              "PH",
              "PK",
              "PL",
              "PT",
              "PW",
              "PY",
              "QA",
              "RO",
              "RS",
              "RU",
              "SA",
              "SB",
              "SC",
              "SL",
              "SD",
              "SE",
              "SG",
              "SH",
              "SI",
              "SK",
              "SN",
              "SO",
              "SR",
              "SS",
              "ST",
              "SV",
              "SX",
              "SY",
              "TC",
              "TF",
              "TG",
              "TH",
              "TJ",
              "TM",
              "TN",
              "TO",
              "TR",
              "TT",
              "TV",
              "TW",
              "TZ",
              "UA",
              "UG",
              "US",
              "UY",
              "UZ",
              "VC",
              "VE",
              "VG",
              "VI",
              "VN",
              "VU",
              "WS",
              "YE",
              "ZA",
              "ZW"
            ]
          },
          "iceClassPolarCode":
          {
            "type": "string",
            "description": "Ice class polar code",
            "enum":
            [
              "PC1",
              "PC2",
              "PC3",
              "PC4",
              "PC5",
              "PC6",
              "PC7",
              "IC",
              "IB",
              "IA",
              "IA_SUPER",
              "NA"
            ]
          },
          "companyNature":
          {
            "type": "string",
            "description": "Nature of the company for the ship referenced in this EMP",
            "enum":
            [
              "SHIPOWNER",
              "ISM_COMPANY"
            ]
          }
        },
        "description": "Ships and emission details"
      },
      "ExternalEmpCarbonCapture":
      {
        "required":
        [
          "captureAndStorageApplied"
        ],
        "type": "object",
        "properties":
        {
          "captureAndStorageApplied":
          {
            "type": "boolean",
            "description": "Are carbon capture and storage technologies being applied. If true, technology and emissionSourceName are required"
          },
          "technology":
          {
            "maxLength": 10000,
            "minLength": 0,
            "type": "string",
            "description": "Technology that is used for carbon capture and storage"
          },
          "emissionSourceName":
          {
            "uniqueItems": true,
            "type": "array",
            "description": "Emission source names this technology is applied to",
            "items":
            {
              "type": "string",
              "description": "Emission source names this technology is applied to"
            }
          }
        },
        "description": "Application of carbon capture and storage technologies"
      },
      "ExternalEmpFuelConsumptionProcedures":
      {
        "required":
        [
          "bunkeringCrossChecksProcedure",
          "equipmentQualityAssuranceProcedure",
          "fuelBunkeredAndInTanksProcedure",
          "informationManagementProcedure",
          "voyagesCompletenessProcedure"
        ],
        "type": "object",
        "properties":
        {
          "fuelBunkeredAndInTanksProcedure":
          {
            "$ref": "#/components/schemas/ExternalEmpProcedureForm"
          },
          "bunkeringCrossChecksProcedure":
          {
            "$ref": "#/components/schemas/ExternalEmpProcedureForm"
          },
          "informationManagementProcedure":
          {
            "$ref": "#/components/schemas/ExternalEmpProcedureForm"
          },
          "equipmentQualityAssuranceProcedure":
          {
            "$ref": "#/components/schemas/ExternalEmpProcedureForm"
          },
          "voyagesCompletenessProcedure":
          {
            "$ref": "#/components/schemas/ExternalEmpProcedureForm"
          }
        },
        "description": "Procedures related to the monitoring of greenhouse gas emissions and fuel consumption"
      },
      "ExternalEmpDataGaps":
      {
        "required":
        [
          "dataSources",
          "fuelConsumptionEstimationMethod",
          "locationOfRecords",
          "responsiblePerson"
        ],
        "type": "object",
        "properties":
        {
          "formulaeUsed":
          {
            "maxLength": 10000,
            "minLength": 0,
            "type": "string"
          },
          "fuelConsumptionEstimationMethod":
          {
            "maxLength": 10000,
            "minLength": 1,
            "type": "string"
          },
          "responsiblePerson":
          {
            "maxLength": 250,
            "minLength": 1,
            "type": "string"
          },
          "dataSources":
          {
            "maxLength": 10000,
            "minLength": 1,
            "type": "string"
          },
          "locationOfRecords":
          {
            "maxLength": 250,
            "minLength": 1,
            "type": "string"
          },
          "itSystem":
          {
            "maxLength": 250,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "ExternalEmpShipEmissions":
      {
        "required":
        [
          "ccsCcu",
          "conditionsOfExemption",
          "emissionsSources",
          "fuelTypes",
          "measuringEquipment",
          "shipDetails",
          "uncertaintyLevel"
        ],
        "type": "object",
        "properties":
        {
          "shipDetails":
          {
            "$ref": "#/components/schemas/ExternalEmpShipDetails"
          },
          "fuelTypes":
          {
            "uniqueItems": true,
            "type": "array",
            "items":
            {
              "$ref": "#/components/schemas/ExternalEmpFuelsAndEmissionsFactors"
            }
          },
          "emissionsSources":
          {
            "uniqueItems": true,
            "type": "array",
            "items":
            {
              "$ref": "#/components/schemas/ExternalEmpEmissionsSources"
            }
          },
          "uncertaintyLevel":
          {
            "uniqueItems": true,
            "type": "array",
            "description": "Level of uncertainty associated with the fuel monitoring methods",
            "items":
            {
              "$ref": "#/components/schemas/ExternalEmpUncertaintyLevel"
            }
          },
          "ccsCcu":
          {
            "$ref": "#/components/schemas/ExternalEmpCarbonCapture"
          },
          "measuringEquipment":
          {
            "uniqueItems": true,
            "type": "array",
            "description": "Description of the measurement instruments involved",
            "items":
            {
              "$ref": "#/components/schemas/ExternalEmpMeasurementDescription"
            }
          },
          "conditionsOfExemption":
          {
            "$ref": "#/components/schemas/ExternalEmpExemptionConditions"
          }
        },
        "description": "Ships and emission details"
      },
      "ExternalEmpRegisteredOwner":
      {
        "required":
        [
          "agreementDate",
          "companyImoNumber",
          "contactName",
          "email",
          "name",
          "ships"
        ],
        "type": "object",
        "properties":
        {
          "name":
          {
            "maxLength": 255,
            "minLength": 1,
            "type": "string",
            "description": "Registered owner name"
          },
          "companyImoNumber":
          {
            "pattern": "^\\d{7}$",
            "type": "string"
          },
          "contactName":
          {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "email":
          {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "agreementDate":
          {
            "type": "string",
            "description": "Date of written agreement (past or present)",
            "format": "date"
          },
          "ships":
          {
            "uniqueItems": true,
            "type": "array",
            "description": "Associated ships",
            "items":
            {
              "$ref": "#/components/schemas/ExternalEmpRegisteredOwnerShipDetails"
            }
          }
        }
      },
      "ExternalEmissionsMonitoringPlan":
      {
        "required":
        [
          "dataGaps",
          "delegatedResponsibility",
          "procedures",
          "shipParticulars"
        ],
        "type": "object",
        "properties":
        {
          "shipParticulars":
          {
            "uniqueItems": true,
            "type": "array",
            "description": "Ships and emission details",
            "items":
            {
              "$ref": "#/components/schemas/ExternalEmpShipEmissions"
            }
          },
          "delegatedResponsibility":
          {
            "$ref": "#/components/schemas/ExternalEmpDelegatedResponsibility"
          },
          "procedures":
          {
            "$ref": "#/components/schemas/ExternalEmpProcedures"
          },
          "dataGaps":
          {
            "$ref": "#/components/schemas/ExternalEmpDataGaps"
          }
        },
        "description": "The emissions monitoring plan data"
      },
...
      "ExternalEmpFuelsAndEmissionsFactors":
      {
        "required":
        [
          "fuelOriginCode",
          "fuelTypeCode",
          "methodDensityBunkerCode",
          "methodDensityTankCode",
          "ttwEFCarbonDioxide",
          "ttwEFMethane",
          "ttwEFNitrousOxide"
        ],
        "type": "object",
        "properties":
        {
          "fuelOriginCode":
          {
            "type": "string",
            "enum":
            [
              "FOSSIL",
              "BIOFUEL",
              "RFNBO"
            ]
          },
          "fuelTypeCode":
          {
            "type": "string",
            "enum":
            [
              "ETHANOL",
              "BIO_DIESEL",
              "HVO",
              "BIO_LNG",
              "BIO_METHANOL",
              "BIO_H2",
              "HFO",
              "LFO",
              "MDO",
              "MGO",
              "LNG",
              "LPG_BUTANE",
              "LPG_PROPANE",
              "H2",
              "NH3",
              "METHANOL",
              "E_DIESEL",
              "E_METHANOL",
              "E_LNG",
              "E_H2",
              "E_NH3",
              "E_LPG",
              "E_DME",
              "OTHER"
            ]
          },
          "otherFuelType":
          {
            "maxLength": 30,
            "minLength": 0,
            "type": "string",
            "description": "Description of the fuel type if not listed (fuelTypeCode OTHER)"
          },
          "ttwEFCarbonDioxide":
          {
            "minimum": 0,
            "type": "number",
            "description": "Tank to wake emission factor for carbonDioxide. Positive or zero decimal number with integer part max 12 digits"
          },
          "ttwEFMethane":
          {
            "minimum": 0,
            "type": "number",
            "description": "Tank to wake emission factor for methane. Positive or zero decimal number with integer part max 12 digits"
          },
          "ttwEFNitrousOxide":
          {
            "minimum": 0,
            "type": "number",
            "description": "Tank to wake emission factor for nitrousOxide. Positive or zero decimal number with integer part max 12 digits"
          },
          "methodDensityBunkerCode":
          {
            "type": "string",
            "enum":
            [
              "ON_BOARD_MEASUREMENT_SYSTEMS",
              "FUEL_SUPPLIER",
              "LABORATORY_TEST",
              "NA"
            ]
          },
          "methodDensityTankCode":
          {
            "type": "string",
            "enum":
            [
              "MEASUREMENT_SYSTEMS",
              "FUEL_SUPPLIER",
              "LABORATORY_TEST",
              "NA"
            ]
          }
        }
      },
      "ExternalEmpRegisteredOwnerShipDetails":
      {
        "required":
        [
          "name",
          "shipImoNumber"
        ],
        "type": "object",
        "properties":
        {
          "shipImoNumber":
          {
            "pattern": "^\\d{7}$",
            "type": "string"
          },
          "name":
          {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          }
        },
        "description": "Associated ships"
      },
      "ExternalEmpManagementProcedures":
      {
        "required":
        [
          "adequacyCheckProcedure",
          "dataFlowActivitiesProcedure",
          "monitoringReportingRoles",
          "riskAssessmentProcedure"
        ],
        "type": "object",
        "properties":
        {
          "monitoringReportingRoles":
          {
            "type": "array",
            "description": "Monitoring and reporting roles",
            "items":
            {
              "$ref": "#/components/schemas/EmpMonitoringReportingRole"
            }
          },
          "adequacyCheckProcedure":
          {
            "$ref": "#/components/schemas/ExternalEmpProcedureForm"
          },
          "dataFlowActivitiesProcedure":
          {
            "$ref": "#/components/schemas/ExternalEmpProcedureForm"
          },
          "riskAssessmentProcedure":
          {
            "$ref": "#/components/schemas/ExternalEmpProcedureForm"
          }
        },
        "description": "Management procedures"
      },
      "ExternalEmpMeasurementDescription":
      {
        "required":
        [
          "emissionSourceName",
          "name"
        ],
        "type": "object",
        "properties":
        {
          "name":
          {
            "maxLength": 250,
            "minLength": 1,
            "type": "string"
          },
          "technicalDescription":
          {
            "maxLength": 10000,
            "minLength": 0,
            "type": "string"
          },
          "emissionSourceName":
          {
            "uniqueItems": true,
            "type": "array",
            "description": "Emission source names this device is used for",
            "items":
            {
              "type": "string",
              "description": "Emission source names this device is used for"
            }
          }
        },
        "description": "Description of the measurement instruments involved"
      },
      "Violation":
      {
        "type": "object",
        "properties":
        {
          "fieldName":
          {
            "type": "string"
          },
          "message":
          {
            "type": "string"
          }
        }
      },
...
      "ExternalEmpExemptionConditions":
      {
        "required":
        [
          "derogationCodeUsed"
        ],
        "type": "object",
        "properties":
        {
          "derogationCodeUsed":
          {
            "type": "boolean",
            "description": "Derogation for monitoring the amount of fuel consumed used. If true, minimumNumberOfVoyages is required"
          },
          "minimumNumberOfVoyages":
          {
            "minimum": 301,
            "type": "integer",
            "format": "int32"
          }
        },
        "description": "Conditions of exemption from per voyage reporting"
      },
      "ExternalEmpFuelOriginTypeName":
      {
        "required":
        [
          "fuelOriginCode",
          "fuelTypeCode"
        ],
        "type": "object",
        "properties":
        {
          "fuelOriginCode":
          {
            "type": "string",
            "enum":
            [
              "FOSSIL",
              "BIOFUEL",
              "RFNBO"
            ]
          },
          "fuelTypeCode":
          {
            "type": "string",
            "enum":
            [
              "ETHANOL",
              "BIO_DIESEL",
              "HVO",
              "BIO_LNG",
              "BIO_METHANOL",
              "BIO_H2",
              "HFO",
              "LFO",
              "MDO",
              "MGO",
              "LNG",
              "LPG_BUTANE",
              "LPG_PROPANE",
              "H2",
              "NH3",
              "METHANOL",
              "E_DIESEL",
              "E_METHANOL",
              "E_LNG",
              "E_H2",
              "E_NH3",
              "E_LPG",
              "E_DME",
              "OTHER"
            ]
          },
          "otherFuelType":
          {
            "maxLength": 30,
            "minLength": 0,
            "type": "string"
          },
          "slipPercentage":
          {
            "maximum": 100,
            "exclusiveMaximum": false,
            "minimum": 0,
            "type": "number",
            "description": "Slip percentage value. Decimal number between 0 and 100 (inclusive) with up to 3 integer digits and 2 fractional digits"
          }
        }
      },
      "ExternalEmpControlActivitiesProcedures":
      {
        "required":
        [
          "correctionsProcedure",
          "dataReviewProcedure",
          "documentationProcedure",
          "outsourcedActivitiesProcedure",
          "qaItProcedure"
        ],
        "type": "object",
        "properties":
        {
          "qaItProcedure":
          {
            "$ref": "#/components/schemas/ExternalEmpProcedureForm"
          },
          "dataReviewProcedure":
          {
            "$ref": "#/components/schemas/ExternalEmpProcedureForm"
          },
          "correctionsProcedure":
          {
            "$ref": "#/components/schemas/ExternalEmpProcedureForm"
          },
          "outsourcedActivitiesProcedure":
          {
            "$ref": "#/components/schemas/ExternalEmpOutsourcedActivitiesProcedure"
          },
          "documentationProcedure":
          {
            "$ref": "#/components/schemas/ExternalEmpProcedureForm"
          }
        },
        "description": "Control Activities procedures"
      },
      "ExternalEmpProcedures":
      {
        "required":
        [
          "controlActivitiesProcedures",
          "emissionsProcedures",
          "fuelConsumptionProcedures",
          "managementProcedures"
        ],
        "type": "object",
        "properties":
        {
          "emissionsProcedures":
          {
            "$ref": "#/components/schemas/ExternalEmpEmissionsProcedures"
          },
          "fuelConsumptionProcedures":
          {
            "$ref": "#/components/schemas/ExternalEmpFuelConsumptionProcedures"
          },
          "managementProcedures":
          {
            "$ref": "#/components/schemas/ExternalEmpManagementProcedures"
          },
          "controlActivitiesProcedures":
          {
            "$ref": "#/components/schemas/ExternalEmpControlActivitiesProcedures"
          }
        }
      },
...
      "ExternalEmpUncertaintyLevel":
      {
        "required":
        [
          "levelOfUncertaintyTypeCode",
          "monitoringMethodCode",
          "shipSpecificUncertainty"
        ],
        "type": "object",
        "properties":
        {
          "monitoringMethodCode":
          {
            "type": "string",
            "enum":
            [
              "BDN",
              "BUNKER_TANK",
              "FLOW_METERS",
              "DIRECT"
            ]
          },
          "levelOfUncertaintyTypeCode":
          {
            "type": "string",
            "enum":
            [
              "DEFAULT",
              "SHIP_SPECIFIC"
            ]
          },
          "shipSpecificUncertainty":
          {
            "maximum": 100,
            "exclusiveMaximum": false,
            "minimum": 0,
            "type": "number",
            "description": "Specific value for the level of uncertainty. Decimal number between 0 and 100 (inclusive) with up to 3 integer digits and 2 fractional digits"
          }
        },
        "description": "Level of uncertainty associated with the fuel monitoring methods"
      },
      "ExternalEmpEmissionsSources":
      {
        "required":
        [
          "emissionSourceClassCode",
          "emissionSourceTypeCode",
          "fuelTypeCodes",
          "monitoringMethods",
          "name"
        ],
        "type": "object",
        "properties":
        {
          "name":
          {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "emissionSourceTypeCode":
          {
            "type": "string",
            "enum":
            [
              "MAIN_ENGINE",
              "AUX_ENGINE",
              "GAS_TURBINE",
              "BOILER",
              "INERT_GAS_GENERATOR",
              "FUEL_CELLS",
              "WASTE_INCINERATOR",
              "OTHER"
            ]
          },
          "emissionSourceClassCode":
          {
            "type": "string",
            "enum":
            [
              "BOILERS",
              "FUEL_CELLS",
              "GAS_TURBINE",
              "ICE",
              "INERT_GAS_GENERATOR",
              "LNG_DIESEL_DFSS",
              "LNG_LBSI",
              "LNG_OTTO_DFMS",
              "LNG_OTTO_DFSS",
              "WASTE_INCINERATORS"
            ]
          },
          "fuelTypeCodes":
          {
            "uniqueItems": true,
            "type": "array",
            "items":
            {
              "$ref": "#/components/schemas/ExternalEmpFuelOriginTypeName"
            }
          },
          "monitoringMethods":
          {
            "uniqueItems": true,
            "type": "array",
            "items":
            {
              "type": "string",
              "enum":
              [
                "BDN",
                "BUNKER_TANK",
                "FLOW_METERS",
                "DIRECT"
              ]
            }
          },
          "identificationNumber":
          {
            "maxLength": 30,
            "minLength": 0,
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes":
    {
      "bearerAuth":
      {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}